+++ pandapower/converter/powerfactory/pf_export_functions.py
@@ -23,6 +23,7 @@ def create_network_dict(app, flag_graphics='GPS'):
         'ElmVac',
         'ElmVsc',
         'ElmVscmono',
+        'ElmSvs',
 
         # branch elements:
         'ElmLne',
@@ -56,6 +57,7 @@ def create_network_dict(app, flag_graphics='GPS'):
         'ElmPvsys': ['W', 'var', 'VA'],
         'ElmXnet': ['W', 'var', 'VA'],
         'ElmSym': ['W', 'var', 'VA'],
+        'ElmSvs': ['W', 'var', 'VA'],
         'ElmAsm': ['W', 'var', 'VA'],
         'ElmShnt': ['W', 'var', 'VA'],
         'ElmZpu': ['W', 'var', 'VA'],